+++ app/src/main/java/com/project981/dundun/view/calendar/CalendarFragment.kt
@@ -1,10 +1,12 @@
-package com.project981.dundun
+package com.project981.dundun.view.calendar
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.project981.dundun.R
+
 class CalendarFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
