+++ app/src/main/java/com/project981/dundun/view/map/MapFragment.kt
@@ -1,16 +1,18 @@
-package com.project981.dundun
+package com.project981.dundun.view.map
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-class HomeFragment : Fragment() {
+import com.project981.dundun.R
+
+class MapFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false)
+        return inflater.inflate(R.layout.fragment_map, container, false)
     }
 }
