+++ app/src/main/java/com/project981/dundun/view/search/SearchFragment.kt
@@ -1,18 +1,18 @@
-package com.project981.dundun
+package com.project981.dundun.view.search
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.project981.dundun.R
 
-class SettingFragment : Fragment() {
+class SearchFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_setting, container, false)
+        return inflater.inflate(R.layout.fragment_search, container, false)
     }
-
 }
