+++ app/src/main/res/navigation/my_nav.xml
@@ -9,27 +9,27 @@
     <!-- 각 fragment에 추가할 수 있는 요소 찾기 -->
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.project981.dundun.HomeFragment"
+        android:name="com.project981.dundun.view.home.HomeFragment"
         android:label="HOME"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.project981.dundun.SearchFragment"
+        android:name="com.project981.dundun.view.search.SearchFragment"
         android:label="search"
         tools:layout="@layout/fragment_search" />
     <fragment
         android:id="@+id/calendarFragment"
-        android:name="com.project981.dundun.CalendarFragment"
+        android:name="com.project981.dundun.view.calendar.CalendarFragment"
         android:label="fragment_calendar"
         tools:layout="@layout/fragment_calendar" />
     <fragment
         android:id="@+id/mapFragment"
-        android:name="com.project981.dundun.MapFragment"
+        android:name="com.project981.dundun.view.map.MapFragment"
         android:label="fragment_map"
         tools:layout="@layout/fragment_map" />
     <fragment
         android:id="@+id/settingFragment"
-        android:name="com.project981.dundun.SettingFragment"
+        android:name="com.project981.dundun.view.setting.SettingFragment"
         android:label="fragment_setting"
         tools:layout="@layout/fragment_setting" />
 </navigation>
